Strip self-linking headings
Some publishing engines—e.g. documentation sites like GitHub README
pages—automatically generate an ID for each heading element and add an
interactive UI element to facilitate copying a URL directly to that heading or
section, typically via a small `<a href="#...">` element that is shown only on
mouse hover.

However, when such pages are added to Readeck, the CSS styling is lost and with
it all nuance about the rendering of such links, leading to SVG icons for
copying header links being shown in Readeck as block elements. Such links are
useless in Readeck because the user viewing their bookmark typically won't need
to permalink to a specific part of that bookmark.

This content pipeline step attempts to strip any such heading links.