Refactored CSV import adapters
This adds a csvBaseAdapter that takes care of the basic process of
reading CSV entries from one or several files (as in the pocket export
file where we read all .csv files from its zip file).

csvAdapter, readwiseAdapter and pocketFileAdapter are now all based
on csvBaseAdapter with their own types and functions.

Tests were refactored to make them easier to write and to cover the
ImportWorker process.

This adds a pkg/csvstruct package that can generate a scanner to read
CSV lines into a struct.