Wallabag import: fix article content originally fetched in non-UTF8 encodings
Wallabag 2.6.13 stores archived page content as UTF-8 regardless of the
original encoding, but for some reason will also store the original
content-type of the page inclusive of the `charset=...` declaration and include
that in API responses. Readeck would then think that the page is still in that
encoding and transcode the already valid content into garbage.

This strips `charset=...` directives from Wallabag.

Co-authored-by: Olivier Meunier <olivier@neokraft.net>