Introducing pkg/http/request
This package provides the needed blocks to add a proper request's URL
to the context, as well as the actual and "real" remote address.

It takes care of allowed proxies. It only changes the request.URL's
scheme and host.

We still change the request.URL to avoid a massive refactor but we leave
other request's property untouched.