Improved image "content negotiation"
Currently, there's no real content negotiation for images in srcset or
source tags, only something based on the optional type attribute.

Part of the conversion process consists of converting picture elements
to img with an srcset attribute.

Before archival, every img[srcset] attribute is reconstructed
in order to sort the list and have the biggest candidates first (and
discard images that will be way too big)

The archiver will then try to download every image in the given order
and, once successful, set an src attribute to the image (with the winner)
and remove the srcset.

Finally, this adds 2 new steps in the CleanDomProcessor:

- remove empty id attributes
- remove img with no src and not srcset attribute