Show full size image in articles
This adds a modal-images controller to the article's content block.
Each image in the content gets a button revealed on hover or focus
that, when clicked, opens a dialog with the image, full width (maximum
100% of screen width).

The dialog (as well as the diagnosis one) triggers a scroll event due
to its fixed position, which propagates to the scroll-progress controller.
To avoid triggering the progression tracker, we skip the event when
the dialog opens.