fftr: honor all elements matched by body selector
As I came to understand, FiveFilters' XPath selectors for isolating an article
body are tried in sequence until something matches. Then, all of the matched
nodes are concatenated together as the new body.

https://help.fivefilters.org/full-text-rss/site-patterns.html#pattern-format
https://forum.fivefilters.org/t/help-creating-site-pattern-need-to-add-image/403/2

This changes the body-extracting logic to preserve all matched nodes for an
XPath selector as long as they have element children.

As a bonus, this also logs the matched XPath selector string.