Embedded video player for extracted videos in articles
Each extracted video receives a "data-iframe-params" attribute that
can be parsed with URLSearchParams. These parameters include the
original source as well as the media's width and height.
From there, we can load the video player in a modal box.

Improved the image modal so we can use real button elements.