Routes to label information using query parameter
A label can be anything and this means you can't always use it as a
path parameter. A label "#/../test" is perfectly valid but, even with
path escaping, it fails in a browser.

We now use the same route ("/labels") for label operations and split
the middlewares and handlers based on the presence of a "name" parameter
in the request URL.

resolves #277 once and for all