Recognize historical, IE-compatible favicon markup
This adds a feature to the favicon extractor to support the following markup:

    <link rel="shortcut icon" href="...">

This markup was historically used for Internet Explorer compatibility, but is
still recognized today by modern browsers and by the HTML5 specification.
However, Readeck used to look for "shortcut-icon" only, which I believe was a
mispelling since I can't find any references to "shortcut-icon" ever being
supported markup in any browser.