Fixed reddit content script
This script doesn't use Reddit's desktop API anymore and only relies
on the JSON version of a post.

From there it can infer the document type and add the necessary content
based on the received data.

The script fully discards the site config file (rendered useless) and
always write a full HTML content.

Thanks @linsui for the initial work!

See #838