Application and Token list in profile views
This introduces a new "Applications" view in the user profile.

An application is nothing more that a token with an associated application.
The "Applications" view retrieves these specific tokens and show them as
applications with details about the registered client.

The token list and token information now only show tokens that have
no associated client.