schema.org metadata extract: article name, description, site name
Previously, this meta tag would result in the article title erroneously getting
assigned to the SiteName property of a bookmark:

    <meta itemprop="name" property="og:title" content="My Article Title" />

First of all, this meta tag combines two different metadata conventions:
schema.org and Open Graph. It's clear from "og:title" that the value supplied
here is meant to be the title of the article. However, Readeck used to assign
the value of "schema.name" to drop.Site, which would in turn get assigned to
Bookmark.SiteName.

With this change:
- "schema.name" is article name
- "schema.description" is article description
- "schema.publisher" is the site name

At the end of the day, per Schema.org specification the `<meta>` tags in document
`<head>` are not valid microdata because they are not part of the visible document.
That means that Readeck observing "schema.name" and "schema.description" is
non-standard, but "schema.author" and "schema.publisher" is according to the
standard (with a caveat that we don't actually verify the "itemtype" property).

https://schema.org/docs/gs.html#microdata_itemscope_itemtype