Remove remote resource URLs that can't be retrieve during archival
In pkg.archiver.processURL, returned errors are mostly errSkippedURL,
which is fine but doesn't provide any control for ignoring and removing
the source (ie. downloading failed).
In that case, we join errSkippedURL and errRemoveSrc that indicates to
remove the source URL entirely so it doesn't stay in the final document.
Resolves #873

On top of that, a final pass now removes the resource elements that have
an empty source URL (src or appropriate attribute).
Resolves #872