make superbus.MemStore.Set behave as with Redis
redis.Client.Set() takes an expiration and would update it if the value
already exists. It wasn't the case with the memory store. So far, there
was no need to update currently stored values with an expiration but it
might soon change.

To do so, we store the timers (by key), and perform any needed update
or removal.