Replaced codeberg.org/readeck/file-compose with an internal tool.
The previous tool was mostly a giant hack around go-jq. It worked but
had limitation and would sadly not keep the object key original order
(and a predictable order matters a bit for the API documentation).

This is only based on go-yaml and deals better with "$merge" entries.
It also introduces a "$include" key so we can have separate markdown
files to the documentation.

Boy do I miss RAML... :)