Remove icons from extracted content
During the final collector, we remove squarish images that are small
enough to be icons.
The test is: max(w, h) / min(w, h) >= 0.9 && max(w, h) <= 64
Every image falling in that range is removed.

This only occurs when Readability is enabled.

In order to do that, archiver.ErrSkippedURL and archiver.ErrRemoveSrc
are now public and ErrRemoveSrc is always joined with ErrSkippedURL.

In bookmarks.ConvertCollectedImage we now always return ErrRemoveSrc
in case of error during the process (which should never happen)

Resolves #854