Use a split token in password recovery
Instead of simply storing an uuid -> user id in the k/v store, we store
the user ID, a code, a verifier and a hash of code+verifier with the key
being the base64url(code).

The recovery code is base64url(code + verifier).
Upon verification, we split code and verifier, retrieve the full
recover code and verify its hash using the code sent to the user.

This avoids timing attacks on the recovery key.
See https://soatok.blog/2021/08/20/lobste-rs-password-reset-vulnerability/