Scope schema.org microdata parsing to just Article-like types
StackExchange is an example of a Q&A site that makes heavy use of schema.org
microdata.  Here is the microdata hierarchy in a typical HTML document from
that site:

- https://schema.org/QAPage
  - https://schema.org/Question
    - author: https://schema.org/Person
    - https://schema.org/Comment
      - author: https://schema.org/Person
    - https://schema.org/Answer
      - author: https://schema.org/Person
      - https://schema.org/Comment
        - author: https://schema.org/Person

Because so many Persons have been found by the current microdata extraction
approach, all of the answer and comment authors on the whole page would get
indiscriminately added to article authors in Readeck.

The new approach only ever extracts names from "publisher" or "author" nodes
that actually apply to Article-like types and not to other types.