Do not save the user after calling User.SetPassword
User.SetPassword saves the row in DB and, since it's always called with
a seed update, it would result in 2 update statements.
This removes the update after SetPassword and performs it when needed.

This commit adds a testing.TestUser.Reset that is called during
some tests teardown (when they trigger a password and/or seed change).