Support saving math generated with MathJax and KaTeX
MathJax v2 outputs HTML+CSS and MathML together. Since we're satisfied with
just MathML, this strips the HTML+CSS part since that would look bad if ever
viewed in Readeck.

MathJax v3 supports CHTML and SVG output modes, but neither outputs MathML.
However, MathJax does expose the original LaTeX expression via the "data-latex"
attribute. Fetch the value of that attribute and use a 3rd-party library to
render LaTeX to MathML when archiving the bookmark.

KaTeX does output MathML, but it also outputs its own HTML-CSS format that is
marked with `aria-hidden="true"`. Ensure that this is stripped even in the
mode when Readability is disabled.