Return an error list when bookmark filters contain errors
Filters are retrieved in withBookmarkListSelectDataset but we can check
if there's a template context so we know if we can return an error
straight away in an API route.
In a view, the form is attached to the template context and errors are
already visible on the result.

Also removed the server.Server pointer in apiRouter, it hasn't been
needed for a long time now.

resolves #975