Moved the authentication providers to internal/server
Authentication providers are very much tied to server operations and
function and having them in internal/auth made it impossible to deal
with the session cookie handler when we need it.

internal/auth depended on the request ID for logging, removed this as
well with a specific logger interface.

Introducing a server.SessionHandler function to retrieve the cookie
handler so we can create cookies from scratch when needed.