Improved dialog styling
They don't need to be in a fixed position with a z-index; the
backdrop pseudo-element already takes care of that.

.dialog is now centered on the page using flexbox on the backdrop with
a default to 100% and 100vh but it can be overridden when needed.

body:has(.dialog:open) would not work on webkit and is replaced by
body:has(.dialog[open]) (same for .dialog-image and .dialog-video)

dialogs always receive the default font (face and size) and a background
and foreground color so they don't inherit from the element they're in.

All dialog classes are components so we can override their properties.